reduce cognitive complexity of memory.write

+++ src/cache/cache.go
@@ -97,7 +97,7 @@ func (c *Cache) Truncate(newSize int64) (memoryDelta int, err error) {
 
 // Write writes data to the cache entry at the specified offset.
 // Returns the memory usage change (negative if memory was freed, positive if more memory was used).
-func (c *Cache) Write(off int64, data bytes, storeInMemory bool, maxMergeSize int) (memoryDelta int, err error) {
+func (c *Cache) Write(off int64, data bytes, storeInMemory bool, maxMergeSize int64) (memoryDelta int, err error) {
 	if len(data) == 0 {
 		return 0, nil // Nothing to write, no memory change
 	}
